requirements.txt
# Core
flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.1.0
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.13.3
lxml==5.3.1

# AI Content Generation
openai==1.68.2

# WordPress Integration
python-wordpress-xmlrpc==2.3

# Scheduling
schedule==1.2.2
APScheduler==3.11.0

# Image handling
Pillow==11.1.0

# Database
peewee==3.17.9

# Utilities
python-dateutil==2.9.0
markdownify==0.14.1
slugify==0.0.1
